requirements.txt

ccxt
pandas
python-telegram-bot==13.15
flask
apscheduler
